+++ _test/test/goldens/generated_build_script.dart
@@ -5,24 +5,24 @@ import 'package:provides_builder/builders.dart' as _i2;
 import 'package:build_test/builder.dart' as _i3;
 import 'package:build_config/build_config.dart' as _i4;
 import 'package:build_modules/builders.dart' as _i5;
-import 'package:build_web_compilers/builders.dart' as _i6;
-import 'package:build/build.dart' as _i7;
-import 'package:build_vm_compilers/builders.dart' as _i8;
+import 'package:build_vm_compilers/builders.dart' as _i6;
+import 'package:build_web_compilers/builders.dart' as _i7;
+import 'package:build/build.dart' as _i8;
 import 'dart:isolate' as _i9;
 import 'package:build_runner/build_runner.dart' as _i10;
 import 'dart:io' as _i11;
 
 final _builders = <_i1.BuilderApplication>[
+  _i1.apply('provides_builder:some_builder', [_i2.someBuilder],
+      _i1.toDependentsOf('provides_builder'),
+      hideOutput: true,
+      appliesBuilders: ['provides_builder:some_post_process_builder']),
   _i1.apply('provides_builder:some_not_applied_builder', [_i2.notApplied],
       _i1.toNoneByDefault(),
       hideOutput: true),
   _i1.apply('provides_builder:throwing_builder', [_i2.throwingBuilder],
       _i1.toDependentsOf('provides_builder'),
       hideOutput: true),
-  _i1.apply('provides_builder:some_builder', [_i2.someBuilder],
-      _i1.toDependentsOf('provides_builder'),
-      hideOutput: true,
-      appliesBuilders: ['provides_builder:some_post_process_builder']),
   _i1.apply(
       'build_test:test_bootstrap',
       [_i3.debugIndexBuilder, _i3.debugTestBuilder, _i3.testBootstrapBuilder],
@@ -35,28 +35,54 @@ final _builders = <_i1.BuilderApplication>[
       hideOutput: true,
       appliesBuilders: ['build_modules:module_cleanup']),
   _i1.apply(
-      'build_web_compilers:ddc_modules',
+      'build_vm_compilers:modules',
+      [_i6.metaModuleBuilder, _i6.metaModuleCleanBuilder, _i6.moduleBuilder],
+      _i1.toNoneByDefault(),
+      isOptional: true,
+      hideOutput: true,
+      appliesBuilders: ['build_modules:module_cleanup']),
+  _i1.apply(
+      'build_vm_compilers:vm', [_i6.vmKernelModuleBuilder], _i1.toAllPackages(),
+      isOptional: true,
+      hideOutput: true,
+      appliesBuilders: ['build_vm_compilers:modules']),
+  _i1.apply('build_vm_compilers:entrypoint', [_i6.vmKernelEntrypointBuilder],
+      _i1.toRoot(),
+      hideOutput: true,
+      defaultGenerateFor: const _i4.InputSet(include: [
+        'bin/**',
+        'tool/**',
+        'test/**.dart.vm_test.dart',
+        'example/**',
+        'benchmark/**'
+      ])),
+  _i1.apply('build_web_compilers:sdk_js_copy', [_i7.sdkJsCopyBuilder],
+      _i1.toNoneByDefault(),
+      hideOutput: true,
+      appliesBuilders: ['build_web_compilers:sdk_js_cleanup']),
+  _i1.apply(
+      'build_web_compilers:dart2js_modules',
       [
-        _i6.ddcMetaModuleBuilder,
-        _i6.ddcMetaModuleCleanBuilder,
-        _i6.ddcModuleBuilder
+        _i7.dart2jsMetaModuleBuilder,
+        _i7.dart2jsMetaModuleCleanBuilder,
+        _i7.dart2jsModuleBuilder
       ],
       _i1.toNoneByDefault(),
       isOptional: true,
       hideOutput: true,
       appliesBuilders: ['build_modules:module_cleanup']),
   _i1.apply(
-      'build_web_compilers:dart2js_modules',
+      'build_web_compilers:ddc_modules',
       [
-        _i6.dart2jsMetaModuleBuilder,
-        _i6.dart2jsMetaModuleCleanBuilder,
-        _i6.dart2jsModuleBuilder
+        _i7.ddcMetaModuleBuilder,
+        _i7.ddcMetaModuleCleanBuilder,
+        _i7.ddcModuleBuilder
       ],
       _i1.toNoneByDefault(),
       isOptional: true,
       hideOutput: true,
       appliesBuilders: ['build_modules:module_cleanup']),
-  _i1.apply('build_web_compilers:ddc', [_i6.ddcKernelBuilder, _i6.ddcBuilder],
+  _i1.apply('build_web_compilers:ddc', [_i7.ddcKernelBuilder, _i7.ddcBuilder],
       _i1.toAllPackages(),
       isOptional: true,
       hideOutput: true,
@@ -65,7 +91,7 @@ final _builders = <_i1.BuilderApplication>[
         'build_web_compilers:dart2js_modules',
         'build_web_compilers:dart_source_cleanup'
       ]),
-  _i1.apply('build_web_compilers:entrypoint', [_i6.webEntrypointBuilder],
+  _i1.apply('build_web_compilers:entrypoint', [_i7.webEntrypointBuilder],
       _i1.toRoot(),
       hideOutput: true,
       defaultGenerateFor: const _i4.InputSet(include: [
@@ -77,53 +103,27 @@ final _builders = <_i1.BuilderApplication>[
         'test/**.node_test.dart',
         'test/**.vm_test.dart'
       ]),
-      defaultOptions: _i7.BuilderOptions({
+      defaultOptions: _i8.BuilderOptions({
         'dart2js_args': ['--minify']
       }),
-      defaultReleaseOptions: _i7.BuilderOptions({'compiler': 'dart2js'}),
+      defaultReleaseOptions: _i8.BuilderOptions({'compiler': 'dart2js'}),
       appliesBuilders: ['build_web_compilers:dart2js_archive_extractor']),
-  _i1.apply('build_web_compilers:sdk_js_copy', [_i6.sdkJsCopyBuilder],
-      _i1.toNoneByDefault(),
-      hideOutput: true,
-      appliesBuilders: ['build_web_compilers:sdk_js_cleanup']),
-  _i1.apply(
-      'build_vm_compilers:modules',
-      [_i8.metaModuleBuilder, _i8.metaModuleCleanBuilder, _i8.moduleBuilder],
-      _i1.toNoneByDefault(),
-      isOptional: true,
-      hideOutput: true,
-      appliesBuilders: ['build_modules:module_cleanup']),
-  _i1.apply(
-      'build_vm_compilers:vm', [_i8.vmKernelModuleBuilder], _i1.toAllPackages(),
-      isOptional: true,
-      hideOutput: true,
-      appliesBuilders: ['build_vm_compilers:modules']),
-  _i1.apply('build_vm_compilers:entrypoint', [_i8.vmKernelEntrypointBuilder],
-      _i1.toRoot(),
-      hideOutput: true,
-      defaultGenerateFor: const _i4.InputSet(include: [
-        'bin/**',
-        'tool/**',
-        'test/**.dart.vm_test.dart',
-        'example/**',
-        'benchmark/**'
-      ])),
   _i1.applyPostProcess(
       'provides_builder:some_post_process_builder', _i2.somePostProcessBuilder,
       defaultGenerateFor: const _i4.InputSet()),
   _i1.applyPostProcess('build_modules:module_cleanup', _i5.moduleCleanup,
       defaultGenerateFor: const _i4.InputSet()),
-  _i1.applyPostProcess(
-      'build_web_compilers:sdk_js_cleanup', _i6.sdkJsCleanupBuilder,
-      defaultReleaseOptions: _i7.BuilderOptions({'enabled': true}),
-      defaultGenerateFor: const _i4.InputSet()),
-  _i1.applyPostProcess(
-      'build_web_compilers:dart_source_cleanup', _i6.dartSourceCleanup,
-      defaultReleaseOptions: _i7.BuilderOptions({'enabled': true}),
-      defaultGenerateFor: const _i4.InputSet()),
   _i1.applyPostProcess('build_web_compilers:dart2js_archive_extractor',
-      _i6.dart2jsArchiveExtractor,
-      defaultReleaseOptions: _i7.BuilderOptions({'filter_outputs': true}),
+      _i7.dart2jsArchiveExtractor,
+      defaultReleaseOptions: _i8.BuilderOptions({'filter_outputs': true}),
+      defaultGenerateFor: const _i4.InputSet()),
+  _i1.applyPostProcess(
+      'build_web_compilers:dart_source_cleanup', _i7.dartSourceCleanup,
+      defaultReleaseOptions: _i8.BuilderOptions({'enabled': true}),
+      defaultGenerateFor: const _i4.InputSet()),
+  _i1.applyPostProcess(
+      'build_web_compilers:sdk_js_cleanup', _i7.sdkJsCleanupBuilder,
+      defaultReleaseOptions: _i8.BuilderOptions({'enabled': true}),
       defaultGenerateFor: const _i4.InputSet())
 ];
 main(List<String> args, [_i9.SendPort sendPort]) async {